backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
openpyxl
python-dotenv